requirements.txt
gensim
openai-whisper
streamlit
pytube
